refactor: simplify store structure and remove unnecessary folders
