+++ src/entities/Category/categoryType.ts
@@ -1,0 +1,8 @@
+import { SubcategoryProps } from '../Subcategory/subcategoryType';
+
+export interface CategoryProps {
+  id: number;
+  title: string,
+  description?: string,
+  subcategory?: SubcategoryProps[],
+}
