+++ src/entities/Subcategory/subcategoryType.ts
@@ -1,0 +1,7 @@
+export interface SubcategoryProps {
+  id: number,
+  title?: string,
+  description?: string,
+  category?: string,
+  category_id?: number,
+}
