+++ src/widgets/Sidebar/Sidebar.tsx
@@ -3,7 +3,7 @@ import classes from './Sidebar.module.scss';
 import { CategoriesList } from '../CategoriesList/CategoriesList';
 
 export const Sidebar = () => {
-  
+
   return (
     <div className={classes.Sidebar}>
       <Typography variant='h3'>
